+++ lib/main.dart
@@ -10,7 +10,7 @@ import 'features/developer/presentation/screens/developer_info_page.dart';
 import 'views/client/client_info_page.dart';
 import 'views/bank/bank_account_page.dart';
 import 'features/profile/presentation/screens/profile_setting_page.dart';
-import 'features/home/presentation/screens/home_page.dart';
+import 'views/home/home_page.dart';
 import 'features/jobs/presentation/screens/website_page.dart';
 import 'features/jobs/presentation/screens/category_details.dart';
 import 'features/shared/models/profile_args.dart';
